Add typescript to project, convert all files to typescript and added interfaces and types
