+++ components/ColorBox.tsx
@@ -1,7 +1,8 @@
 import React from 'react';
 import { View, Text, StyleSheet } from 'react-native';
+import { IColor } from '../interfaces/interfaces';
 
-const ColorBox = ({ colorName, hexCode }) => {
+const ColorBox = ({ colorName, hexCode }: IColor) => {
   const boxColor = {
     backgroundColor: hexCode,
   };
