+++ interfaces/interfaces.ts
@@ -1,0 +1,14 @@
+export interface IColor {
+  colorName: string;
+  hexCode: string;
+}
+
+export interface IColorPalette {
+  paletteName: string;
+  colors: IColor[];
+}
+
+export interface IPalettePreview {
+  handlePress: () => void;
+  colorPalette: IColorPalette;
+}
