+++ package.json
@@ -25,8 +25,13 @@
   "devDependencies": {
     "@babel/core": "^7.8.6",
     "@react-native-community/eslint-config": "^2.0.0",
+    "@types/jest": "^26.0.0",
+    "@types/react": "^16.9.36",
+    "@types/react-native": "^0.62.13",
+    "@types/react-test-renderer": "^16.9.2",
     "babel-preset-expo": "~8.1.0",
-    "eslint": "^7.2.0"
+    "eslint": "^7.2.0",
+    "typescript": "^3.9.5"
   },
   "private": true
 }
